Move into identity package
